garbage_classification/requirements.txt
torch==2.0.1+cpu
torchvision==0.15.2+cpu
Flask==2.3.2
Pillow==9.3.0
